[김규민] 로또 2단계 (수동) (#1220)
* refactor: `StringSplitter` 객체 추가

입력값을 split하는 책임을 수행하는 객체 추가, 위임

* test: `StringSplitterTest` 추가

* feat: `LottoNumber` 생성

int형의 로또 번호를 갖는 `LottoNumber` 생성

* feat: `LottoNumbers` 및 관련 테스트 추가

* feat: 당첨 통계를 결정하는 `WinnerResult` 객체 생성

(미완성)

* feat: 구매한 로또와 당첨번호를 비교하여 결과를 얻어내는 기능

`LottoResult` 등 추가

* feat: 당첨 통계를 계산하고 출력하는 기능

* test: 테스트 추가

GoldenTicketTest, LottoResultTest, LottoTicketTest, WinnerStatisticsTest

* refactor: rename `StringSplitter` to `StringParser`

* refactor: 1단계 구현 완료

* refactor: `LottoNumber` 인스턴스 캐싱

총 45개의 `LottoNumber` 객체만 생성되도록 수정

* refactor: `PurchaseResult` 생성

티켓 오피스에서 티켓 발급 후 리턴할 객체

* feat: 로또 수동 구매 기능 구현

* refactor: rename `LottoTicketOffice` to `LottoTicketMachine`

* refactor: `StringParser` delimiter 추가

* refactor: `Rank`에 `getValuesAscendingOrder()` 메서드 추가

* refactor: 수동 구입전략 테스트 삭제

* test: @BeforeEach로 중복 코드 정리

* refactor: LottoTicket 생성 시 validation(중복 요소) 추가

* refactor: 메서드 추출, 위치 조정

* refactor: `LottoNumber` getter 삭제

* test: LottoNumbersPicker의 결과가 오름차순인지 테스트

* refactor: `LottoJudge`, `Rank` 알고리즘 수정

* `Rank.of()`에 count와 containsBonusBall을 넘기도록 수정
* 1등, 2등의 count값을 현실과 동일한 값으로 수정

* refactor: `Rank` 내 금액 필드의 타입을 BigDecimal로 변경

* refactor: `PurchaseAmount` 내 금액값 필드의 타입을 BigDecimal로 변경

* refactor: extract method(`LottoController`)

* refactor: 모든 매개변수에 final 추가

* refactor: `ExceptionMessages` 추가, 예외 메시지에 활용

* refactor: `Rank` 테스트 추가

* refactor: `PurchaseAmount` 테스트 추가

* refactor: 테스트 클래스 formatting

class 접근 제어자의 public 삭제

* refactor: `Rank`.`matchCount`의 타입을 Integer로 변경

* NONE인 경우 `matchCount`를 null로 수정

* refactor: reformat `LottoController`

* refactor: `Rank.of()` 메서드 수정

* refactor: `PurchaseAmount` 생성 시 validation 추가

+ 관련 테스트 추가

* refactor: 수동으로 구매한 티켓도 출력하도록 수정 (`ConsoleOutputView`)

* refactor: `Rank` 오름차순 정렬 방식 수정

+++ src/test/java/lotto/domain/AutoPurchaseTest.java
@@ -3,6 +3,7 @@ package lotto.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -14,7 +15,7 @@ class AutoPurchaseTest {
     void purchase() {
         // given
         LottoPurchase lottoPurchase = new AutoPurchase();
-        PurchaseAmount amount = new PurchaseAmount(14_000);
+        PurchaseAmount amount = new PurchaseAmount(BigDecimal.valueOf(14_000));
 
         // when
         List<LottoTicket> lottoTickets = lottoPurchase.purchase(amount);
